requirements.txt
black==24.10.0
djangorestframework==3.15.2
fastapi[all]==0.115.8
fastapi-utils==0.8.0
flake8==7.1.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1 
isort==5.13.2
mypy==1.14.1
pylint==3.3.4
SQLAlchemy==2.0.38
typing-inspect==0.9.0
typing_extensions==4.12.2
